fix(tests): wait for license to be applied before trying to login (#8744)
once the zero cluster is up, it applies a trial license. Until the license is
not applied, health check for alpha returns that the cluster is an OSS cluster.

This PR ensures that the t driver code now waits for the license to be
proposed and applied in the cluster. Then, it waits for the login to
succeed before running the tests.