chore: wait longer for login to work (#8816)
Seems like it takes longer for a 6 alpha cluster to be available
for running queries. This PR also makes the WaitForRestore
function more robust. It is possible that we try to get alpha's
health while it is in draining mode and we need to keep retrying
until it is ready to server again.