chore(ci): add workflow for OSS build + unit tests (#8834)
## Problem

Previous commits broke OSS build. Since we didn't have a test for this,
it was undetected by CI.

## Solution

Add a CI job to verify that we can make the dgraph OSS build, and run
all dgraph OSS unit tests. We do this in a separate workflow because
this does not require a separate self-hosted runner (we can use github
hosted runners for this job).

See also: https://github.com/dgraph-io/dgraph/pull/8832