chore: waitForRestore should check for receiving snapshot (#8851)
when we do a restore, it is possible that one of the alphas do not do a
restore but instead just stream the snapshot from one of the other
alpahs. In the dgraphtest package, in the WaitForRestore function, we
should wait for either restore to complete or snapshot to be streamed.