chore: run multitenancy test on all combos (#8852)
This PR runs multi-tenancy upgrade tests for more combinations of Dgraph
versions. Additionally, it performs some code cleanups too. It renames
suite variable to an appropriate name so that the package name `suite`
does not conflict with variable name. It also adds more Dgraph versions
to the list for running upgrade test against.