chore(ci): fix test coverage workflow (#8864)
## Problem

On merges to main, coverage cannot be sent because workflow is
attempting to checkout a merge commit, which is not applicable on a push
to main (or a scheduled run).

## Solution

Add conditional logic to workflow. Unfortunately AFAIK it is not
possible to add the conditional logic to the specific step in the job,
so we have to split this into two jobs, and select which one based on
the type of event. This creates some duplication but seems unavoidable.