dgraphtest: add workaround for multiple groot issue in export-import
When we perform an export and import into the cluster, this ends
up creating two groot users in the cluster. This PR implements
a workaround so that two groot users are not created to begin with.
With this workaround, now we can also run upgrade tests using
ExportImport upgrade strategy.