fix(bulk): enable running bulk loader with only gql schema
If only gql schema is provided, and DQL schema is missing
bulk loader used to fail. This PR adds the ability to
auto generate the DQL schema from the gql schema and use
it in the bulk loader.

This PR also adds support for running bulk loader in the
dgraphtest package.