chore(upgrade): add workaround for multiple groot issue in export-import (#8897)
When we perform an export and import into the cluster, this ends up
creating two groot users in the cluster. This PR implements a workaround
so that two groot users are not created to begin with. With this
workaround, now we can also run upgrade tests using ExportImport.

The workaround is as follows. We find the UIDs of groot user and
guardian group from the new cluster. We modify the output of the export
and replace the UIDs of groot user and guardian group with the UIDs we
found in the new cluster.

Closes: DGRAPHCORE-300