fix(upgrade): look for version string in logs bottom up
For in place upgrade, the container remains the same and has
the version string twice, one for the old container and one
for the new container. We want the version string in the logs
for the new container. Hence, we look for the string in the logs
bottom up by using strings.LastIndex instead of strings.Index.