feat(query): add feature flag normalize-compatibility-mode (#8845) (#8929)
This feature flag allows configuring @normalize in the query language
such that if there are multiple fields with same alias, they could be
returned as a list or non-list.

There was a breaking change at some point with the PR
https://github.com/dgraph-io/dgraph/pull/7639. Some customers do not
prefer a list response in case of @normalize. We would like to not make
any breaking changes going forward, and hence, introducing this feature
flag enables users to make that choice.