feat(debug): add parse_key to debug tool (#7640)
This adds a `dgraph debug --parse_key` flag that can return the ParsedKey
struct of a hex key. This is useful if there's a lone key that you want parse
without having to have a p directory around. This flag does not need a p
directory, just the hex key string.

Example:

$ dgraph debug --parse_key
000000000000000000000b6467726170682e74797065000000000000000001
     {d} Key: UID: 1, Attr: 0-dgraph.type, Data key

This tells you that the key
000000000000000000000b6467726170682e74797065000000000000000001
is for the predicate `0-dgraph.type` and the UID `1`.

Docs PR: https://github.com/dgraph-io/dgraph-docs/pull/636

Co-authored-by: Aman Mangal <aman@dgraph.io>