feat(dql): Add Dot Product Op Precedence. (#7)
Description: Adds definition of dot product into operator precedence for
the "math dql" operation. Also adds unit tests for same.
At the same time, this addresses an issue for the operator precedence
specified by the multiply and divide operators. These were previously
given with the precedence of "/" being higher than "*", but this does
not follow the standard left-right with left-binding parsing
requirement! Added unit test for the same
Docs: No PR for this as of yet, but this is definitely a change in
behavior!!