fix(dql): Repair a broken type assertion error. (#15)
Description: There was a broken Type assertion and value assignment in
worker/mutation.go that is strongly suspected to be the cause of CICD
errors.

An oddity here: the change that I made looks totally unnecessary! It
looks like the code should be either:
edge.Value = b.Value.([]byte)
or
edge.Value = *(b.Value.(*[]byte))

Yet, for some reason, it seems that both results are coming out!
When I copy the code snippets to a go-playground instance, it only
allows the second option, but it seems that most of the tests break on
the second option, and some break on the first option. I do not have an
explanation for the behavior, but I know that the code works.

Any reviewers, I would appreciate your insights!

Closes: DGR-65