fix(restore): allow incrementalFrom to be 1 in restore API (#8988)
This is a minor inconvenience while using the incremental restore API
that when incrementalFrom is set to 1, we throw an error back. The
restore API takes two backup numbers incrementalFrom & backupNum and
restores all the backups including both the ends, i.e. following set
notation all the backups in the set [incrementalFrom, backupNum] are
restored. This should work fine when incrementalFrom is set to 1 which
is a full backup.