Indexing(Vector): Adding several temporary changes for the alpha release (#38)
Description:
Tracing is temporarily being disabled as it seems to slow down index
mutations substantially. This is being done for alpha release. This
needs to be re-enabled while avoiding the slowdown caused by
construction string messsages from mutation objects/requests.

The second change is to not update the schema by explicitly adding
"<vector_pred>__vector_" predicate to the schema. This increases
transaction overhead of building/updating the index.

The third change is a hack to not delete the HNSW predicate when zero
recommends it.

Fixes: DGR-236