feat(indexing): Adopting GenericOpts and templated VectorIndex. (#59)
Description: Pulls in changes from vector-indexer that allow us to
specify configuration options when creating a VectorIndex. This will
allow us (for example) to specify the metric type for the vector index
(euclidian vs dotproduct vs cosine) separate from the technology for the
index type (hnsw vs vamana vs kdtree vs hypertree).

Note that at the moment, the process still retains everything using
float64 arrays, even though we want to switch to float32. We will take
this in two steps to keep the relative sizes of the changes a bit more
reasonable.

Closes: Hyp-14