Bug(GraphQL): Fixed HYP-336 (#70)
Description: Made the embedding parameeter, by, to be NonNull for
`querySimilar<Type>ById `& `querySimilar<Type>ByEmbedding ` queries.
This fixes the backend panic.

However, the postman client fails to show the dropdown as in the input
for the parameter. It shows it as a simple text input box. The client,
while editing the query text in the editor pane shows valid enum values
as part of the autocomplete feature.

Fixes: HYP-336