Feat(GraphQL): Add support for @search directive for vector searches (#69)
Description: This PR adds support for specifying @search directive with
search options (vector index options)
The fix is not backward compatible. Earlier searchArgs in @search
directive were expected to be Enum of index types.
Arguments are expected to be strings now of the following form.
      `@search(by: [String!])`
where each string element in the array is of the form
      `<searchArg> := <searchType> [  ( <searchOptions> ) ]`
searchOptions are optional.
       ```
```
<searchOptions> :=  <searchOption> , <searchOption>, ...
       <searchOption> :=  <optionName><COLON><SPACE><optionValue>
```

```
For example:

`product_vector: [Float!] @hm_embedding @search(by: ["hnsw(metric: euclidian, exponent: 6)"])`

Fixes: HYP-313