fix(core): fix deadlock in runMutation and txn.Update() (#9085)
We call txn.Update() once the transaction has finished. This moves all
the uncommited posting lists to a delta map in the txn. This can cause a
deadlock with runMutation. Mostly we have seen it when a transaction has
a timeout. RunMutation keeps on going, while txn.Update() gets
triggered. This PR fixes it by not calling txn.Update() if the
transaction has failed. We would futher look into cancelling runMutation
at that time too.

Fixes:
https://linear.app/hypermode/issue/DGR-477/dgraph-v24-alpha-2-hangs

---------

Co-authored-by: ShivajiKharse <115525374+shivaji-dgraph@users.noreply.github.com>
Co-authored-by: shivaji-dgraph <shivaji@dgraph.io>