perf(query): Read just the latest value for scalar types (#8966)
Description: Right now, when we read some data, we read the entire
posting list. With this we populate List structure in memory, which has
two layers. These two layers are then merge sorted. All of this takes a
lot of time. In this diff, we are updating the query processor to just
read the latest value in case of scalar types.

LDBC test for IC09.
Before: 39 seconds
After: 30 seconds