refactor code for mutable layer and fix bugs in count index (#9199)
Currently the delta postings are stored in a map. When we use the cache,
these deltas gets unnecessarily copied again and again because of the
limitations of the map. We have introduced a new struct for holding the
mutable layer, that would allow us to work on the list of deltas. 

With the introduction of this, we can now pass around posting lists from
one transaction to another without copying the map.

We also store results such as delete markers, uid to posting map, length
to improve query performance.

This PR also fixes bugs in the count index introduced in the PR #9208.