fix(core): Improve error handling in hnsw module (#9236)
This PR improves error handling in the hnsw module. In some cases not
all errors have to be propagated to the next caller, in this case the
error string is currently compared with `strings.Contains(err.Error(),
"some_error_text")` or directly `err.Error() == "some_error_string"`.

This is bad practice. Instead errors should be directly compared with
the native `errors.Is(err, target)` function, which traverses the tree
of each error and detects if it contains the specified target error.

This refactor also fixes two occurrences where the caller of a function
compared the error string to a value which could never occur.