feat(graphql): output translated DQL query when the graphql debug superflag is set (#9280)
**Description**

This change adds the translated DQL query to the GraphQL's `extension`
attribute set when the GraphQL debug superflag is set to true.

In Dgraph, GraphQL queries are translated to DQL before entering the
query system. This change allows for an inspection of what the
translator is doing behind the scenes.

Potential Use Cases:

* learning DQL
* aiding few shot prompting in LMs
* use in Dgraph "dashboards" or other helper apps

Note: this is for queries (reads) only. Mutations in DQL are pretty
straightforward, but support for them could be added in the future.


![image](https://github.com/user-attachments/assets/a818ef04-5189-4bc1-a3d7-7b7a5d2fbe70)

**Checklist**

- [x] Code compiles correctly and linting passes locally
- [x] For all _code_ changes, an entry added to the `CHANGELOG.md` file
describing and linking to
      this PR
- [x] Tests added for new functionality, or regression tests for bug
fixes added as applicable
- [x] [Docs PR](https://github.com/dgraph-io/dgraph-docs/pull/689)