chore: update minor deps to latest working versions (#9294)
This PR upgrades all indirect dependencies to the latest versions that I
have tested to work without other code changes. With this, go.mod is
broadly at the latest and greatest for everything.

There are a number of deprecated libraries still in here, such as all
the opencensus dependencies, but upgrading them to the opentelemetry
equivalents is a larger change that can be done separately.