Reduce memory usage during Alter execution
We currently use y.Throttle type to ensure that
we do not concurrently build indexes for many
predicates at the same time to avoid opening
too many files on disk. This type allocates large
memory initially to achieve this. By default, on
my machine we allocate 128 MB of memory every time
Alter is called. Now, Go is slow to release this
memory back to OS increasing the memory usage.

This PR sets the max to 1024 predicates at a time.
It would be lower if Alter request involves fewer
predicates in the request