fix(core): Add maxfrontiersize to shortest path query (#9382)
Currently we use dijkstra's algorithm to find shortest path. Sometimes
the heap could grow indefinitely, which would slow down the query
significantly and cause dgraph to go out of memory. This would only
happen in very specific instances depending upon the graph. To fix this,
we are introducing a new argument, maxfrontiersize in shortest path
query. This would limit the size of the heap fixing the issue.

New query would look like:

```
{ 
  path as shortest(from: 0x2, to: 0x5, numpaths:5, maxfrontiersize: 10000) {
  connected_to
    @facets(weight)
 }
```

Fixes: https://github.com/hypermodeinc/dgraph/issues/9333