feat(mcp): Add variable support to MCP run_query tool (#9464)
Add support for variables to the MCP query tool.
Accept 'variables' argument.
variables must be a JSON map of parameters used in the query string when
using a parametrized query.

**Description**

Add variables argument to query tool
use QueryWithVars instead of Query and pass var map to QueryWithVars

**Checklist**

- [x] Code compiles correctly and linting passes locally
- [ ] For all _code_ changes, an entry added to the `CHANGELOG.md` file
describing and linking to
      this PR
- [x] Tests added for new functionality, or regression tests for bug
fixes added as applicable
- [ ] For public APIs, new features, etc., PR on
[docs repo](https://github.com/hypermodeinc/docs) staged and linked here

---------

Co-authored-by: mattthew <matthew.mcneely@gmail.com>