fix(query): return full float value in query results (#9492)
**Description**

Updates `float` formatting to express full values. Switches from `%f` to
`%g`, which according to [Golang docs](https://pkg.go.dev/fmt) will use
as many digits as necessary to express the value.

>for %g [the default precision] is the smallest number of digits
necessary to identify the value uniquely.

Closes #9491.

**Discussion**

This changes the output format but not semantics of float values in
result data. Currently floats have a fixed precision of 6, filled with
trailing zeros as necessary.

```json
{ "float": 1.000000 }
```

The new specifier `%g` drops unnecessary trailing zeros.

```json
{ "float": 1 }
```

I think this OK as they are semantically equivalent. Since unnecessary
characters are dropped, it also saves bytes on the wire. It also aligns
with the `bigfloat` format, which never gives trailing zeros.

**Checklist**

- [x] Code compiles correctly and linting passes locally
- [x] For all _code_ changes, an entry added to the `CHANGELOG.md` file
describing and linking to this PR
- [x] Tests added for new functionality, or regression tests for bug
fixes added as applicable